elsar: Functions and Tools For Creating ELSA Data Stacks

Provides tools and data pipelines to support Essential Life Support Areas (ELSA) conservation planning. Includes functions to load and harmonize global spatial datasets, prepare raster layers for spatial prioritization, create zone layers (protect, restore, manage), and calculate coverage metrics over planning units.

Ver también

Autor-a

Maintainer: Scott Atkinson scott.atkinson@undp.org (ORCID)

Authors:

Sandra Neubert s.neubert@uq.edu.au (ORCID)