Function to create a static plot from raster data

elsar_plot_static_raster() allows plotting both continuous and discrete data from a SpatRaster in ggplot or tmap. It can be combined with elsar_plot_optics() and elsar_plot_extra_data() to create reproducible plots.

Usage

elsar_plot_static_raster(
  raster_in,
  data_type = c("continuous", "discrete"),
  type = "ggplot_vector",
  categorical = FALSE,
  number_categories = 10,
  data_layer = NULL,
  background = NULL,
  extend_background = 0.05,
  custom_palette = NULL,
  plot_title = "",
  legend_title = NULL,
  color_map = "viridis",
  expand_plot = FALSE,
  raster_df_out = FALSE
)

Arguments

raster_in

The SpatRaster file to be plotted.

data_type

Character. Either "continuous" (default) or "discrete" to specify how data should be displayed.

type

A character of the plot type. Either "ggplot_vector", "ggplot_raster" or "tmap".

categorical

Logical. Only used when data_type = "discrete". If TRUE, data is already categorical.

number_categories

Integer. Only used when data_type = "discrete". Number of categories to split continuous data into. Default is 10.

data_layer

Character. Only used when data_type = "discrete". The data layer name for categorization.

background

Requires a SpatRaster input (preferably with the same data as raster_in) to plot the background data.

extend_background

A numerical value that allows to extent the background beyond the extent of raster_in. If extend_background <= 1, the lat and lon extend will be extended by the ratio provided (e.g. 0.05 will extend it by 5%). If extend_background > 1 all sides will be extended by the absolute value provided.

custom_palette

An optional custom palette for plotting. Default uses the viridis package.

plot_title

An optional plot title.

legend_title

An optional legend title.

color_map

The name of the viridis palette to be used. Default is "viridis".

expand_plot

Logical. Whether to expand the plot, so there is no space between border and plot.

raster_df_out

Logical. Whether to return only the plot (FALSE) or also the data frame behind the plot (TRUE). Needed for elsar_plot_distribution().

Value

A ggplot or tmap object, or a list with the plot and the data frame if raster_df_out = TRUE.

Examples

boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd"
)
#> [2025-12-19 19:19:26] Creating boundary for planning region...
#> [2025-12-19 19:19:26] Creating custom Mollweide projection centred on planning region...

pus <- make_planning_units(
  boundary_proj = boundary_proj,
  pu_size = NULL,
  pu_threshold = 8.5e5,
  limit_to_mainland = FALSE
)
#> Error in eval(assertion, env): argument "iso3" is missing, with no default
wad_dat <- get_wad_data()
#> [2025-12-19 19:19:26] Loading WAD subset data from package...
#> [2025-12-19 19:19:26] WAD subset data loaded successfully.

wadOut <- make_normalised_raster(
  raster_in = wad_dat,
  pus = pus,
  iso3 = "NPL"
)
#> Error: object 'pus' not found

# Continuous data plot
(wad_plot <- elsar_plot_static_raster(
  raster_in = wadOut,
  type = "ggplot_raster",
  background = wad_dat,
  legend_title = "wad"
))
#> Error: object 'wadOut' not found

# Discrete data plot
(wad_plot_d <- elsar_plot_static_raster(
  raster_in = wadOut,
  data_type = "discrete",
  type = "ggplot_raster",
  number_categories = 10,
  background = wad_dat,
  data_layer = "wad_final_cog"
))
#> Error: object 'wadOut' not found