Calculate Areal Coverage of a Zone Relative to Planning Units

Calculates percentage of a zone layer relative to the total area of planning units.

Uso

get_coverage(zone_layer, pu_layer)

Argumentos

zone_layer

A binary SpatRaster representing a target zone.

pu_layer

A binary SpatRaster representing planning units.

Valor

Numeric. Percentage of zone coverage across PUs.