Get Specific LULC Class Value

Retrieves a specific class value for a given LULC product. This is the primary interface for consumer functions to obtain class values dynamically based on the LULC product being used.

Usage

get_lulc_class_value(product, class_name)

Arguments

product

Character. LULC product ID: one of "esri_10m", "dynamic_world", or "esa_worldcover".

class_name

Character. Standardized class name to retrieve. Common names include: "agriculture", "built_area", "trees", "water", "forest_managed", "forest_all".

Value

Integer or integer vector of class values for the specified class.

Examples

# Get agriculture class value for different products
get_lulc_class_value("esri_10m", "agriculture")       # Returns 5
#> [1] 5
get_lulc_class_value("dynamic_world", "agriculture")  # Returns 4
#> [1] 4
get_lulc_class_value("esa_worldcover", "agriculture") # Returns 40
#> [1] 40

# Get built area class value
get_lulc_class_value("esri_10m", "built_area")        # Returns 7
#> [1] 7
get_lulc_class_value("dynamic_world", "built_area")   # Returns 6
#> [1] 6

# Get managed forest classes (may return vector)
get_lulc_class_value("esri_10m", "forest_managed")    # Returns c(31, 32, 40, 53)
#> [1] 31 32 40 53