Create Buffered Polygons from Point Geometries

This function creates polygon buffers around point geometries within a given spatial layer. It uses the specified area attribute to determine the size of each buffer and transforms geometries into a specified coordinate reference system for accurate area calculations.

Usage

convert_points_polygon(
  wdpa_layer,
  area_attr = "REP_AREA",
  area_crs = "ESRI:54009",
  nQuadSegs = 50,
  append_sf = TRUE
)

Arguments

wdpa_layer

An sf object containing point geometries and the area attribute.

area_attr

A character string specifying the name of the area attribute. Default is 'REP_AREA'.

area_crs

A character string specifying the coordinate reference system for area calculations. Default is 'ESRI:54009' (World Mollweide projection).

nQuadSegs

An integer specifying the number of segments to use for buffering. Default is 50.

append_sf

logical. If TRUE, returns the initial wdpa_layer with polygons instead of points. If FALSE, will only returned the points with buffers.

Value

An sf object with buffered polygon geometries.

Examples

if (FALSE) { # \dontrun{
 pa <- wdpar::wdpa_fetch("NPL",
  wait = TRUE,
  download_dir = here::here()
)
buffered_polygons <- convert_points_polygon(wdpa_layer = pa,
      area_crs = sf::st_crs(pa))
      } # }