Function to create basic histogram with categorical data

Function to create basic histogram with categorical data

Usage

elsar_plot_distribution(
  type = "ggplot",
  categorical = TRUE,
  raster_df = NULL,
  col_interest = NULL,
  bin_number = 10,
  custom_palette = NULL,
  show_legend = FALSE,
  color_map = "viridis"
)

Arguments

type

A character denoting whether "ggplot" or "tmap" is being used.

categorical

logical. Whether data is categorical (TRUE) or continuous (FALSE).

raster_df

A df object of the data to include in the histogram.

col_interest

A character of the column of interest. Uses "interval" if nothing is provided.

bin_number

An integer value of how many bins to split the data into (default = 10).

custom_palette

An optional custom palette for plotting. Default uses the viridis package.

show_legend

logical. Whether to include legend in plot or not. Default is FALSE.

color_map

The name of the viridis palette to be used. Default is "viridis".

Value

A list of a ggplot or tmap object and a SpatRaster with the new background data.