Function to extract general raster data

make_normalised_raster()allows you to align, normalise and save a raster file. This is applicable for the following data This is applicable for the following data sets: certified forests, drought risk, flood risk, intact wilderness area, soc difference, threatened species richness, voc, wad convergence evidence

Usage

make_normalised_raster(
  raster_in,
  pus,
  iso3,
  invert = FALSE,
  rescaled = TRUE,
  name_out,
  output_path = NULL
)

Arguments

raster_in

A SpatRaster file that contains the data to be put into right format

pus

A SpatRaster file that contains the reference spatial extent, crs etc.in form of the planning units

iso3

A string of the iso3 name of the data (country name)

invert

Logical. Default is FALSE. If TRUE, highest values in the original dataset should be valued the lowest in the prioritisation.

rescaled

Logical. Default is TRUE. Should remain TRUE unless there is a very specific reason not to rescale.

name_out

A string with the data name that will be used for the output tiffile

output_path

An optional output path for the created file.

Value

A SpatRaster file that has been aligned and normalised

Examples

boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd",
  do_project = TRUE
)

pus <- make_planning_units(boundary_proj = boundary_proj,
                           pu_size = NULL,
                           pu_threshold = 8.5e5,
                           limit_to_mainland = FALSE)
#> [1] "The current number of planning units is: 3743382"
#> [1] "The current number of planning units is: 2396805"
#> [1] "The current number of planning units is: 1665187"
#> [1] "The current number of planning units is: 1223941"
#> [1] "The current number of planning units is: 741036"
wad_dat <- get_wad_data()

wadOut <- make_normalised_raster(raster_in = wad_dat,
                                 pus = pus,
                                 iso3 = "NPL")