Function to create planning units and custom projection

This function creates planning units for a spatial prioritisation problem in a raster format. It also creates a custom projection (wkt file) centred on the planning region based on the Mollweide projection.

Usage

make_planning_units(
  boundary_proj,
  pu_size = NULL,
  pu_threshold = 850000,
  limit_to_mainland = FALSE,
  iso3,
  output_path = NULL
)

Arguments

boundary_proj

A sfobject representing the boundary of the planning region in the preferred projection. Preferably generated with make_boundary()

pu_size

A way to define a custom planning unit size. Can be NULL to use default settings that generate planning units as small as possible whilst still being computationally efficient.

pu_threshold

An integer value that gives a maximum number of PUs. The default (8.5e5) is set based on prioritizr processing time, network transfer time and solver time.

limit_to_mainland

A logical that determines whether planning units should only be created for mainland area (FALSE; default) or not (TRUE)

iso3

The iso3 country code (character) of the country of interest.

output_path

An optional output path for the created file.

Value

A raster (.tif) file with the planning unit information for the chosen planning region.

Examples

boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd",
  do_project = TRUE
)

pus <- make_planning_units(boundary_proj = boundary_proj,
                           pu_size = NULL,
                           pu_threshold = 8.5e5,
                           limit_to_mainland = FALSE)
#> [1] "The current number of planning units is: 3743382"
#> [1] "The current number of planning units is: 2396805"
#> [1] "The current number of planning units is: 1665187"
#> [1] "The current number of planning units is: 1223941"
#> [1] "The current number of planning units is: 741036"