Function to create existing protected area raster file.

Function to create existing protected area raster file.

Usage

make_protected_areas(
  from_wdpa = TRUE,
  iso3,
  download_path = NULL,
  sf_in,
  status = c("Established", "Inscribed", "Designated"),
  pa_def = 1,
  designation_mab = FALSE,
  buffer_points = TRUE,
  area_column = "REP_AREA",
  nQuadSegs = 50,
  return_sf = FALSE,
  pus,
  output_path = NULL
)

Arguments

from_wdpa

logical. If TRUE, downloads data from the wdpar package. For more information on the data or package, please consult the online package documentation.

iso3

A string of the iso3 name of the data (country name)

download_path

A path that the wdpa data is saved to.

sf_in

An sf object and alternative to downloading the data from the wdpar package. Only needed when from_wdpa is FALSE.

status

A vector containing which status of protected area should be included. Based on the STATUS field of the wdpa database. Default is c("Established", "Inscribed", "Designated").

pa_def

A value or list of values containing which pa definition (1 = protected area; 0 = OECM (not supported yet)) to include. Default is 1.

designation_mab

logical. If FALSE, excludes UNESCO MAB areas.

buffer_points

logical. Only relevant when "POINT" or "MULTIPOINT" geometries exist in the data. If TRUE, creates a circular buffer around "POINT" data based on area information data that is then used as polygon data needed for exactextractr calculations.

area_column

A string of the column name with the area information needed for buffer calculations.

nQuadSegs

An integer specifying the number of segments to use for buffering. Default is 50.

return_sf

logical. Allows to return sfobject if needed. Default is FALSE.

pus

A raster file that contains the reference spatial extent, crs etc.in form of the planning units.

output_path

An optional output path for the created file.

Value

A rasterfile that has been aligned to the planning units

Examples

if (FALSE) { # \dontrun{
boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd"
)

pus <- make_planning_units(
  boundary_proj = boundary_proj,
  pu_size = NULL,
  pu_threshold = 8.5e5,
  limit_to_mainland = FALSE
)

current_pas <- make_protected_areas(
  iso3 = "NPL",
  download_path = here::here(),
  buffer_points = TRUE,
  pus = pus
)
} # }