Function to create protection zone for ELSA analysis

The function creates a SpatRaster used to filter the planning units available for the "protect" action in the ELSA framework based on HFP, agricultural areas and buildings. It can also be used to create a layer used as a locked-out constraint for the same zone.

Usage

make_protection_zone(
  current_pas = NULL,
  hfp_in,
  crop_in = NULL,
  built_in = NULL,
  pus,
  iso3,
  hfp_threshold = 0.95,
  filter_patch_size = TRUE,
  min_patch_size = 20,
  make_locked_out = FALSE,
  output_path = NULL
)

Arguments

current_pas

Either NULL or a vector file with the current protected areas. If NULL, PA sf object will be created with the make_protected_areas() function.

hfp_in

A SpatRaster of the human footprint index. Will be cropped and re-projected with the PUs.

crop_in

A SpatRaster of the crop data. Will be cropped (CURRENTLY: rescaled) and re-projected with the PUs.

built_in

A SpatRaster of building data. Will be cropped (CURRENTLY: rescaled) and re-projected with the PUs.

pus

A SpatRaster file that contains the reference spatial extent, crs etc.in form of the planning units.

iso3

A string of the iso3 name of the data (country name).

hfp_threshold

If <1, HFP threshold is calculated based on a quantile approach (HFP inside PAs excluding x% highest; default is 0.95). If >1, that value will be used as the threshold.

filter_patch_size

Logical. Whether to filter out clumps smaller than a given threshold.

min_patch_size

Positive integer. Clumps smaller than this are removed.

make_locked_out

Logical. Whether to change data from available planning units for a zone to a layer for a locked-out constraint in prioritizr. Will swap 0s and 1s.

output_path

An optional output path for the created file.

Value

A SpatRaster file that either serves as available planning units for the protection zone in the ELSA analysis or locked-out areas for the same zone.

Examples

if (FALSE) {
boundary_proj <- make_boundary(
  boundary_in = boundary_dat,
  iso3 = "NPL",
  iso3_column = "iso3cd",
  do_project = TRUE
)

pus <- make_planning_units(
  boundary_proj = boundary_proj,
  pu_size = NULL,
  pu_threshold = 8.5e5,
  limit_to_mainland = FALSE
)

path_in <- "YourDataPath"

list_dat <- extract_filename_filetype(
  data_name = "hfp",
  file_path = path_in
)

load_hfp <- elsar_load_data(
  file_name = paste0(list_dat["filename"][[1]], list_dat["filetype"][[1]]),
  file_type = list_dat["filetype"][[1]], file_path = path_in
)

list_dat <- extract_filename_filetype(
  data_name = "crop_suit",
  file_path = path_in
)

load_crop <- elsar_load_data(
  file_name = paste0(list_dat["filename"][[1]], list_dat["filetype"][[1]]),
  file_type = list_dat["filetype"][[1]], file_path = path_in
)

list_dat <- extract_filename_filetype(
  data_name = "build",
  file_path = path_in
)

load_built <- elsar_load_data(
  file_name = paste0(list_dat["filename"][[1]], list_dat["filetype"][[1]]),
  file_type = list_dat["filetype"][[1]], file_path = path_in
)

protection_zone1 <- make_protection_zone(
  hfp_in = load_hfp,
  crop_in = load_crop,
  built_in = load_built,
  pus = pus,
  iso3 = "NPL",
  make_locked_out = FALSE
)

protection_zone2 <- make_protection_zone(
  hfp_in = load_hfp,
  crop_in = load_crop,
  built_in = load_built,
  filter_patch_size = FALSE,
  pus = pus,
  iso3 = "NPL",
  make_locked_out = FALSE
)

protection_zone3 <- make_protection_zone(
  hfp_in = load_hfp,
  built_in = load_built,
  hfp_threshold = 13,
  pus = pus,
  iso3 = "NPL",
  make_locked_out = TRUE
)
}