Create Urban Greening Opportunities Raster

This function generates an urban greening opportunities raster by integrating NDVI (Normalized Difference Vegetation Index), land use/land cover (LULC) classification, and urban heat intensity from SDEI/GHS data. The process includes normalizing NDVI, identifying urban areas from LULC, rasterizing urban heat exposure, and combining all layers to map priority areas for urban greening.

Usage

make_urban_greening_opportunities(
  ndvi_raster,
  lulc_raster,
  sdei_statistics,
  pus,
  iso3,
  return_urban_areas = FALSE,
  output_path = NULL,
  cores = 4
)

Arguments

ndvi_raster

A SpatRaster representing NDVI data (vegetation greenness).

lulc_raster

A SpatRaster representing land use/land cover (LULC) classes.

sdei_statistics

An sf object of urban heat exposure statistics (e.g., WBGT).

pus

A SpatRaster defining planning units (PUs) to align outputs to.

iso3

ISO3 country code used for filtering the urban heat data.

return_urban_areas

Logical. If TRUE, also returns the raster of binary urban areas.

output_path

Optional. A directory path to write the resulting raster(s).

cores

Integer. Number of CPU cores to use in multi-threaded operations (if supported). Default = 4.

Value

A SpatRaster of normalized urban greening opportunities. If return_urban_areas = TRUE, a raster stack is returned with both greening opportunities and urban extent.

Examples

if (FALSE) { # \dontrun{
ndvi <- elsar::elsar_load_data(
  file_path = ".",
  file_name = "ndvi.tif"
  )

lulc <- elsar::elsar_load_data(
  file_path = ".",
  file_name = "lulc.tif",
  )

sdei <- sf::read_sf("path_to_sdei_data.gpkg")

result <- make_urban_greening_opportunities(
  ndvi_raster = ndvi,
  lulc_raster = lulc,
  sdei_statistics = sdei,
  pus = planning_units,
  iso3 = "NPL"
)
} # }